requirements.txt
joblib
Pillow
requests
streamlit
streamlit==1.12.0
pandas
altair==4.0